+++ src/Core/GUI/guiCoreV4.java
@@ -6,6 +6,9 @@ import Core.GUI.SwingEX.*;
 import Core.Player.SaveDirectoryConstants;
 import Core.Player.playerData;
 import Core.SFX.audioRepository;
+import Core.events.EMouseListener;
+import Core.events.eventBuilder;
+import Core.events.eventCoreV2;
 import org.omg.PortableInterceptor.SYSTEM_EXCEPTION;
 
 import javax.swing.*;
@@ -3204,34 +3207,92 @@ public class guiCoreV4 {
 
     /** Event Windows **/
 
-    private void loadPopupView(String name, BufferedImage gfx, String desc) { //pop-up window, no effects
-        //TODO: Add in GFX & Buttons.
+    private void loadEventWindow(eventBuilder event) { //loads the event window
 
-        XPanel pnlPopup = new XPanel();
-        layers.add(pnlPopup, new Integer(15), 0);
-        pnlPopup.setBounds((window.getWidth() / 2) - 318, 150, 635, 500);
-        pnlPopup.setVisible(true);
+        try { //because i'm using such a vague declaration, gotta use a try/catch in case something foreign enters
+            event.loadOptions();
 
-        XLabel lblHeader = new XLabel(gfxRepository.greenHeader); //header at the top of the event window
-        pnlPopup.add(lblHeader);
-        lblHeader.setBounds(0, 0, 635, 25);
-        lblHeader.setVisible(true);
+            XPanel pnlPopup = new XPanel();
+            layers.add(pnlPopup, new Integer(15), 0);
+            pnlPopup.setBounds((window.getWidth() / 2) - 318, 150, 635, 500);
+            pnlPopup.setVisible(true);
 
-        XLabel lblPopupBG = new XLabel();
-        pnlPopup.add(lblPopupBG);
-        lblPopupBG.setBounds(0, 25, pnlPopup.getWidth(), pnlPopup.getHeight());
-        lblPopupBG.scaleImage(gfxRepository.menuBackground);
-        lblPopupBG.setVisible(true);
+            XLabel lblHeader = new XLabel(gameSettings.eventhandler.getHeader(event.getType())); //header at the top of the event window
+            pnlPopup.add(lblHeader);
+            lblHeader.setBounds(0, 0, 635, 25);
+            lblHeader.setVisible(true);
 
-        XLabel lblEvtName = new XLabel(name, gfxRepository.txtButtonSmall, gfxRepository.clrText);
-        lblHeader.add(lblEvtName);
-        lblEvtName.setBounds(0, 0, lblHeader.getWidth(), lblHeader.getHeight());
-        lblEvtName.setVisible(true);
+            XLabel lblPopupBG = new XLabel();
+            pnlPopup.add(lblPopupBG);
+            lblPopupBG.setBounds(0, 25, pnlPopup.getWidth(), pnlPopup.getHeight());
+            lblPopupBG.scaleImage(gfxRepository.menuBackground);
+            lblPopupBG.setVisible(true);
 
-        XLabel lblEvtDesc = new XLabel("<html>" + desc + "</html>", gfxRepository.txtButtonSmall, gfxRepository.clrText);
-        lblPopupBG.add(lblEvtDesc);
-        lblEvtDesc.setBounds(15, 15, lblPopupBG.getWidth() - 30, 300);
-        lblEvtDesc.setVisible(true);
+            XLabel lblEvtName = new XLabel(event.getName(), gfxRepository.txtButtonSmall, gfxRepository.clrText);
+            lblHeader.add(lblEvtName);
+            lblEvtName.setBounds(0, 0, lblHeader.getWidth(), lblHeader.getHeight());
+            lblEvtName.setVisible(true);
+
+            XLabel lblEvtDesc = new XLabel("<html>" + event.getDesc() + "</html>", gfxRepository.txtButtonSmall, gfxRepository.clrText);
+            lblPopupBG.add(lblEvtDesc);
+            lblEvtDesc.setBounds(15, 15, lblPopupBG.getWidth() - 30, 200);
+            lblEvtDesc.setVisible(true);
+
+            //TODO: Make more efficient.
+
+            XListSorter eventButtons = new XListSorter(XConstants.VERTICAL_SORT, 5, 20, 220);
+
+            for (int i = 0; i < event.button.size(); i++) { //load in the buttons
+                XButtonCustom btnOption = new XButtonCustom(null, SwingConstants.LEFT);
+                btnOption.setText(event.button.get(i).getButtonText(), gfxRepository.txtSubtitle, gfxRepository.clrText);
+                btnOption.setToolTipText(event.button.get(i).getMouseOverText());
+                btnOption.setPreferredSize(new Dimension(lblPopupBG.getWidth() - 40, 40));
+
+                btnOption.addMouseListener(new EMouseListener(event.button.get(i)) {
+                    XButtonCustom source;
+
+                    @Override
+                    public void mouseClicked(MouseEvent mouseEvent) {
+                        source = (XButtonCustom)mouseEvent.getSource();
+                        source.getContent().setHorizontalAlignment(SwingConstants.RIGHT);
+                        getButton().clickButton();
+                        audioRepository.buttonConfirm();
+
+                    }
+
+                    @Override
+                    public void mousePressed(MouseEvent mouseEvent) {
+                        source = (XButtonCustom)mouseEvent.getSource();
+                        source.getContent().setHorizontalAlignment(SwingConstants.RIGHT);
+                    }
+
+                    @Override
+                    public void mouseReleased(MouseEvent mouseEvent) {
+                        source = (XButtonCustom)mouseEvent.getSource();
+                        source.getContent().setHorizontalAlignment(SwingConstants.LEFT);
+                    }
+
+                    @Override
+                    public void mouseEntered(MouseEvent mouseEvent) {
+                        source = (XButtonCustom)mouseEvent.getSource();
+                        source.getContent().setHorizontalAlignment(SwingConstants.CENTER);
+                    }
+
+                    @Override
+                    public void mouseExited(MouseEvent mouseEvent) {
+                        source = (XButtonCustom)mouseEvent.getSource();
+                        source.getContent().setHorizontalAlignment(SwingConstants.LEFT);
+                    }
+                });
+
+                eventButtons.addItem(btnOption);
+            }
+
+            eventButtons.placeItems(lblPopupBG);
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
 
     }